+++ src/Models/SectionAsset.php
@@ -26,12 +26,6 @@ class SectionAsset extends Asset
     }
 
     #[Override]
-    public function getPermissionName(): string
-    {
-        return Entry::AUTH_ENTRY;
-    }
-
-    #[Override]
     public function getAdminType(): string
     {
         return Yii::t('media', 'ASSET_SECTION_ASSET');